[#6] error handling refactoring
