+++ pom.xml
@@ -24,7 +24,6 @@
 		<java.version>1.8</java.version>
 		<cglib.version>3.2.5</cglib.version>
 		<guava.version>21.0</guava.version>
-		<spring-orm.version>4.2.5.RELEASE</spring-orm.version>
 		<lombok.version>1.16.16</lombok.version>
 		<spock.version>1.1-groovy-2.4-rc-4</spock.version>
         <commons-io.version>2.5</commons-io.version>
@@ -54,12 +53,6 @@
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
 			<version>${guava.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-orm</artifactId>
-			<version>${spring-orm.version}</version>
 		</dependency>
 
 		<!--Test-->
