+++ src/main/java/com/beerjournal/infrastructure/error/BeerJournalException.java
@@ -1,0 +1,10 @@
+package com.beerjournal.infrastructure.error;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class BeerJournalException extends RuntimeException {
+    private final ErrorInfo info;
+}
